Add mission-specific BMG file
