+++ PulsarEngine/UI/UI.cpp
@@ -369,128 +369,6 @@ void ChangeImage(LayoutUIControl &control, const char *paneName, const char *tpl
     }
 };
 
-// Implements the use of Pulsar's BMGHolder when needed
-enum BMGType {
-    BMG_NORMAL,
-    CUSTOM_BMG,
-};
-BMGType isCustom;
-
-static int GetMsgIdxByBmgId(const BMGHolder &bmg, s32 bmgId) {
-    if (bmg.bmgFile == nullptr || bmg.messageIds == nullptr) return -1;
-    const BMGMessageIds &msgIds = *bmg.messageIds;
-    int ret = -1;
-    for (int i = 0; i < msgIds.msgCount; ++i) {
-        int curBmgId = msgIds.messageIds[i];
-        if (curBmgId == bmgId) {
-            ret = i;
-            break;
-        } else if (curBmgId > bmgId)
-            break;
-    }
-    return ret;
-}
-
-static const BMGHolder *matchedCustomBmg = nullptr;
-
-static const BMGHolder *GetCharaNameBmg() {
-    static BMGHolder charaNameBmg;
-    static const void *loadedFile = nullptr;
-
-    ArchiveMgr *archiveMgr = ArchiveMgr::sInstance;
-    if (archiveMgr == nullptr) return nullptr;
-
-    void *file = archiveMgr->GetFile(ARCHIVE_HOLDER_UI, "message/CharaName.bmg", nullptr);
-    if (file == nullptr) {
-        loadedFile = nullptr;
-        charaNameBmg.bmgFile = nullptr;
-        return nullptr;
-    }
-
-    if (file != loadedFile) {
-        charaNameBmg.Init(*reinterpret_cast<const BMGHeader *>(file));
-        loadedFile = file;
-    }
-    return &charaNameBmg;
-}
-
-static int GetMsgIdxById(const BMGHolder &normalHolder, s32 bmgId) {
-    int ret = GetMsgIdxByBmgId(System::sInstance->GetBMG(), bmgId);
-    if (ret >= 0) {
-        isCustom = CUSTOM_BMG;
-        matchedCustomBmg = &System::sInstance->GetBMG();
-        return ret;
-    }
-    ret = GetMsgIdxByBmgId(System::sInstance->GetBMGCT(), bmgId);
-    if (ret >= 0) {
-        isCustom = CUSTOM_BMG;
-        matchedCustomBmg = &System::sInstance->GetBMGCT();
-        return ret;
-    }
-    ret = GetMsgIdxByBmgId(System::sInstance->GetBMGBT(), bmgId);
-    if (ret >= 0) {
-        isCustom = CUSTOM_BMG;
-        matchedCustomBmg = &System::sInstance->GetBMGBT();
-        return ret;
-    }
-    const BMGHolder *charaNameBmg = GetCharaNameBmg();
-    if (charaNameBmg != nullptr) {
-        ret = GetMsgIdxByBmgId(*charaNameBmg, bmgId);
-        if (ret >= 0) {
-            isCustom = CUSTOM_BMG;
-            matchedCustomBmg = charaNameBmg;
-            return ret;
-        }
-    }
-    isCustom = BMG_NORMAL;
-    matchedCustomBmg = nullptr;
-    ret = GetMsgIdxByBmgId(normalHolder, bmgId);
-    return ret;
-}
-kmBranch(0x805f8c88, GetMsgIdxById);
-
-wchar_t *GetMsgByMsgIdx(const BMGHolder &bmg, s32 msgIdx) {
-    const BMGInfo &info = *bmg.info;
-    if (msgIdx < 0 || msgIdx >= info.msgCount) return nullptr;
-    const u32 offset = info.entries[msgIdx].dat1Offset & 0xFFFFFFFE;
-    const BMGData &data = *bmg.data;
-    return reinterpret_cast<wchar_t *>((u8 *)&data + offset);
-}
-
-wchar_t *GetMsg(const BMGHolder &normalHolder, s32 msgIdx) {
-    wchar_t *ret = nullptr;
-    if (isCustom == CUSTOM_BMG && matchedCustomBmg != nullptr) {
-        ret = GetMsgByMsgIdx(*matchedCustomBmg, msgIdx);
-    }
-    if (ret == nullptr) ret = GetMsgByMsgIdx(normalHolder, msgIdx);
-    return ret;
-}
-kmBranch(0x805f8cf0, GetMsg);
-
-const u8 *GetFontIndex(const BMGHolder &bmg, s32 msgIdx) {
-    const BMGInfo &info = *bmg.info;
-    if (msgIdx < 0 || msgIdx >= info.msgCount) return nullptr;
-    return &info.entries[msgIdx].font;
-};
-
-const u8 *GetFont(const BMGHolder &normalHolder, s32 msgIdx) {
-    const u8 *ret = nullptr;
-    if (isCustom == CUSTOM_BMG && matchedCustomBmg != nullptr) {
-        ret = GetFontIndex(*matchedCustomBmg, msgIdx);
-    }
-    if (ret == nullptr) ret = GetFontIndex(normalHolder, msgIdx);
-    return ret;
-}
-kmBranch(0x805f8d2c, GetFont);
-
-const wchar_t *GetCustomMsg(s32 bmgId) {
-    const BMGHolder &bmg = System::sInstance->GetBMG();
-    int msgIdx = GetMsgIdxById(bmg, bmgId);
-    if (isCustom == CUSTOM_BMG && matchedCustomBmg != nullptr) {
-        return GetMsgByMsgIdx(*matchedCustomBmg, msgIdx);
-    }
-    return GetMsgByMsgIdx(bmg, msgIdx);
-}
 void ResetMatColor(lyt::Pane *pane, u32 color) {
     lyt::Material *mat = pane->material;
     ut::Color colors(color);
